add swiper to main components
